test: Move utils.ts to test folder
